add: gameModet.test

+++ js/controller/gameController.js
@@ -27,6 +27,12 @@ class GameController {
         }
 
         this.gameInterval = setInterval(() => {
+            console.log(this.model.gameOverFlag);
+            if (this.model.gameOverFlag) {
+                clearInterval(this.gameInterval);
+                alert('Game Over');
+                return;
+            }
             this.model.drop();
             this.view.render(this.model.grid, this.model.currentTetromino);
         }, this.interval);
